+++ CHANGELOG.md
@@ -11,6 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [0.1.8] - 2025-04-17
+
+### Add
+
+- Possibility to key bind the commands to a keyboard shortcut.
+- Automated integration definition file sanity check.
+- Preview of the documentation and a link to the full online documentation on some keywords.
+
 ## [0.1.7] - 2025-04-16
 
 ### Updated